+++ LochieGraphics/Sync.cpp
@@ -206,7 +206,7 @@ bool Sync::Update(
 			timeSinceStep = 0.0f;
 			float randPercentage = std::rand() / (float)RAND_MAX;
 
-			int index = floor(randPercentage * 30.0f);
+			int index = (int)floor(randPercentage * 30.0f);
 			SceneManager::scene->audio.PlaySound((Audio::SoundIndex)(index + (int)Audio::syncFootSteps00));
 		}
 		fireDirection = move;
